Deleted the old react-dapp and renamed client to "react-dapp"
