+++ examples/react-dapp/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react-dapp",
+  "name": "client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -8,13 +8,14 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "@types/jest": "^24.9.1",
-    "@types/node": "^12.12.57",
-    "@types/react": "^16.9.49",
+    "@types/node": "^12.19.1",
+    "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
+    "axios": "^0.20.0",
     "iam-client-lib": "git+https://github.com/energywebfoundation/iam-client-lib.git#develop",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-scripts": "3.4.3",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-scripts": "4.0.1",
     "typescript": "^3.7.5"
   },
   "scripts": {
@@ -37,5 +38,6 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  }
+  },
+  "proxy": "http://localhost:3333/"
 }
