+++ examples/react-dapp/src/App.tsx
@@ -1,80 +1,70 @@
-import React, { useEffect, useRef, useState } from "react";
-import "./App.css";
-import { IAM } from "iam-client-lib";
-import { Button, CircularProgress, Typography } from "@material-ui/core";
+import React, { useState } from 'react'
+import './App.css'
+import { IAM, CacheServerClient } from 'iam-client-lib'
+import axios from 'axios'
+import Button from '@material-ui/core/Button'
+import { Typography } from '@material-ui/core'
+
+const cacheClient = new CacheServerClient({ url: 'https://volta-iam-cacheserver.energyweb.org/' })
+
+const iam = new IAM({
+  rpcUrl: 'https://volta-rpc.energyweb.org',
+  chainId: 73799,
+  cacheClient,
+})
+
+type Role = {
+  name: string
+  namespace: string
+}
+
+type User = {
+  did: string
+  verifiedRoles: Role[]
+}
 
 function App() {
-  const iam = useRef<IAM | null>(null);
-  const [isLoggedIn, setIsLoggedIn] = useState<boolean>(false);
-  const [userDid, setUserDid] = useState<string>("");
-  const [loggingInProgress, setLoggingInProgress] = useState<boolean>(false);
-
-  useEffect(() => {
-    iam.current = new IAM({
-      rpcUrl: "https://volta-rpc.energyweb.org/",
-      chainId: 73799
-    });
-
-    async function init() {
-      setLoggingInProgress(true);
-      const { connected = false, did, userClosedModal } = (await iam.current?.initializeConnection()) || {};
-      if (userClosedModal) {
-        return window.location.reload()
-      }
-      localStorage.setItem("connected", connected.toString());
-      setIsLoggedIn(connected || false);
-      setUserDid(did || "");
-      setLoggingInProgress(false);
-    }
-
-    const alreadyLoggedIn = localStorage.getItem("connected");
-    if (alreadyLoggedIn) {
-      init();
-    }
-  }, []);
+  const [roles, setRoles] = useState<Role[]>([])
 
   const handleLogin = async () => {
-    setLoggingInProgress(true);
-    const { connected = false, did, userClosedModal } = (await iam.current?.initializeConnection()) || {};
-    if (userClosedModal) {
-      return window.location.reload()
+    const { did } = await iam.initializeConnection({useMetamaskExtension: false})
+    if (!did) {
+      console.log("unable to retrieve DID")
+      return
     }
-    localStorage.setItem("connected", connected.toString());
-    setIsLoggedIn(connected || false);
-    setUserDid(did || "");
-    setLoggingInProgress(false);
-  };
+    const signer = iam.getSigner()
+    const latestBlock = await signer?.provider?.getBlockNumber()
+    const claim = await iam.createPublicClaim({
+      data: { blockNumber: latestBlock },
+    })
+    const { data } = await axios.post<{ token: string }>('/login', {
+      claim,
+    })
+    const config = {
+      headers: { Authorization: `Bearer ${data.token}` },
+    }
+    const { data: roles } = await axios.get<Role[]>('/roles', config)
+    const { data: user } = await axios.get<User>('/user', config)
 
-  const handleLogout = async () => {
-    await iam.current?.closeConnection()
-    localStorage.removeItem('connected');
-    window.location.reload();
+    setRoles(roles)
   }
-
   return (
     <div className="App">
       <header className="App-header">
-        {isLoggedIn ? (
-          <div>
-            <p>Your did:</p>
-            <p>{userDid}</p>
-            <Button onClick={handleLogout} color="secondary" variant="contained">Log out</Button>
-          </div>
-        ) : loggingInProgress ? (
-          <div>
-            <CircularProgress color="secondary" />
-            <Typography>Loading...</Typography>
-          </div>
-        ) : (
-          <div>
-            <Button color="primary" variant="contained" onClick={handleLogin}>
-              Login
-            </Button>
-          </div>
-        )}
+        {roles && roles.length > 0 ? (
+          <>
+            <Typography>User roles:</Typography>
+            {roles.map(({ name, namespace }) => (
+              <p key={namespace}>{`${name} at ${namespace}`}</p>
+            ))}
+          </>
+        ) : null}
+        <Button variant="contained" color="primary" onClick={handleLogin}>
+          login
+        </Button>
       </header>
     </div>
-  );
+  )
 }
 
-export default App;
+export default App
